pyproject.toml
[tool.poetry]
name = "iiwa-setup"
version = "0.1.0"
description = "iiwa real world setup"
authors = ["Nicholas Pfaff <nepfaff@mit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
# drake = { ersion = ">=0.0.20250118 <1.0", source = "drake-nightly" }
drake = ">=1.26.0"
numpy = "^2.0.0"
manipulation = "^2025.3.3"
tqdm = "^4.66.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
ipython = "^8.15.0"
pyyaml = "^6.0.1"
scipy = "<1.17"
termcolor = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
black = "^23.12.1"
isort = "^5.12.0"
poetry-pre-commit-plugin = "^0.1.2"

[[tool.poetry.source]]
name = "drake-nightly"
url = "https://drake-packages.csail.mit.edu/whl/nightly/"
priority = "supplemental"

[tool.isort]
profile = 'black'
lines_between_types = 1
combine_as_imports = true
known_first_party = ['iiwa_setup']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
